requirements.txt
psycopg2-binary==2.9.9
toml==0.10.2
google-generativeai==0.8.2
streamlit==1.37.0
plotly==5.24.1
pandas==2.2.3